ticha-django-site/requirements.txt
Django==1.9.6
Whoosh==2.7.4
beautifulsoup4==4.4.1
certifi==2017.4.17
chardet==3.0.4
django-haystack==2.5.dev1
idna==2.5
lxml==3.6.0
psycopg2==2.6.2
requests==2.18.1
urllib3==1.21.1
